Ensure vim & go are updated
Small fix for the case where go & vim have already been installed with
the scripts, but you run it a second time for updates. There was an
error when cloning the repositories with git, because the folder already
existed. So now we check if the folder exist, clone if not, then cd
& pull to update.